requirements.txt
backoff==2.2.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
Flask-Cors==4.0.0
gunicorn==20.0.4
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
python-dotenv==1.0.0
pytube==15.0.0
requests==2.31.0
urllib3==2.1.0
videodb==0.0.2
Werkzeug==3.0.1
zipp==3.17.0
